Fixed the reference representation to 64bits inside the CUDA kernel "the column bug"

+++ resources/myers_gpu/coreMyersGPU-CC20.cu
@@ -95,7 +95,7 @@ inline __device__ uint32_t funnelShiftL_CC20(const uint32_t currentCandidateEntr
 
 __device__ void myerslocalFermiKernel_CC20( const d_qryEntry_t *d_queries, const uint32_t * d_reference, const bpm_gpu_cand_info_t *d_candidates,
 											const uint32_t *d_reorderBuffer, bpm_gpu_res_entry_t *d_reorderResults, const bpm_gpu_qry_info_t *d_qinfo,
-								 			const uint32_t idCandidate, const uint32_t sizeRef, const uint32_t numReorderedResults, 
+								 			const uint32_t idCandidate, const uint64_t sizeRef, const uint32_t numReorderedResults, 
 											const uint32_t intraQueryThreadIdx, const uint32_t threadsPerQuery)
 {
 	if (idCandidate < numReorderedResults){
@@ -234,7 +234,7 @@ __device__ void myerslocalFermiKernel_CC20( const d_qryEntry_t *d_queries, const
 }
 
 __global__ void myersFermiKernel_CC20(const d_qryEntry_t *d_queries, const uint32_t * d_reference, const bpm_gpu_cand_info_t *d_candidates, const uint32_t *d_reorderBuffer,
-						    		  bpm_gpu_res_entry_t *d_reorderResults, const bpm_gpu_qry_info_t *d_qinfo, const uint32_t sizeRef,  const uint32_t numReorderedResults,
+						    		  bpm_gpu_res_entry_t *d_reorderResults, const bpm_gpu_qry_info_t *d_qinfo, const uint64_t sizeRef,  const uint32_t numReorderedResults,
 						    		  uint32_t *d_initPosPerBucket, uint32_t *d_initWarpPerBucket, uint32_t numWarps)
 {
 	uint32_t bucketIdx = 0;
